requirements.txt
asgiref==3.5.0
cloudinary==1.29.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
gunicorn==20.1.0
phonenumbers==8.12.47
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
